+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Next release
 
 * Keeping attribute prefix while copying XML streams. [#23](https://github.com/difi/vefa-peppol/issues/23)
+* Updating model to allow for multiple process identifiers as part of one process metadata.
 
 
 ## 1.0.1